Feature: Add Ability to Retry Unreceived Items
Add a "Success" flag to the ItemReceivedEventArgs, which can be updated to false by a game client if it knows the item couldn't be received. ReceiveItems() then interprets that to know if it should not mark that item as received, and stop receiving items.

The game client is then responsible for calling ReceiveReady() once it is able to receive items again.